delete make promise biasa bisa, tp pake try catch masih belum paham betul

+++ server/controllers/TodoController.js
@@ -86,21 +86,42 @@ class TodoController {
     }
   }
 
-  static async delete(req, res) {
-    try {
-      const data = await Todo.destroy({
-        id: +req.params.id
-      });
-      res.status(200).json({ data, msg: "todo success to delete" });
-      // if (id) {
-      // } else {
-      //   throw error;
-      // }
-    } catch (error) {
-      res.status(404).json({ msg: "data not found" });
-      res.status(500).json({ msg: "Internal Server Error" });
-    }
+  static delete(req, res) {
+    const id = +req.params.id
+      Todo.findByPk(id)    
+        .then(() => { 
+          if (!id) {
+            res.status(404).json({ msg: "data not found" });
+          } else {
+            Todo.destroy({
+              where: {
+                id: id
+              }
+            })
+            res.status(200).json({ msg: "todo success to delete" })
+          }
+        }) 
+        .catch((err) => { 
+          res.status(500).json({ msg: "Internal Server Error" });
+        })
   }
+
+  // static async delete(req, res) {
+  //   try {
+  //     const id = +req.params.id
+  //     const data = await Todo.destroy({
+  //       id: id
+  //     });
+  //     res.status(200).json({ data, msg: "todo success to delete" });
+  //     // if (id) {
+  //     // } else {
+  //     //   throw error;
+  //     // }
+  //   } catch (error) {
+  //     res.status(404).json({ msg: "data not found" });
+  //     res.status(500).json({ msg: "Internal Server Error" });
+  //   }
+  // }
 }
 
 module.exports = TodoController;
